Refactor EditPostModal
