+++ packages/ui/src/forum/components/PostList/PostEditor.tsx
@@ -1,12 +1,16 @@
+import { createType } from '@joystream/types'
 import React, { useMemo, useState } from 'react'
 import styled from 'styled-components'
 
 import { ButtonGhost, ButtonPrimary, ButtonsGroup } from '@/common/components/buttons'
 import { CKEditor } from '@/common/components/CKEditor'
+import { useApi } from '@/common/hooks/useApi'
 import { useModal } from '@/common/hooks/useModal'
 import { PostListItemType } from '@/forum/components/PostList/PostListItem'
+import { useForumPostParents } from '@/forum/hooks/useForumPostParents'
 import { EditPostModalCall } from '@/forum/modals/PostActionModal/EditPostModal'
 import { ForumPost } from '@/forum/types'
+import { useProposalPostParents } from '@/proposals/hooks/useProposalPostParents'
 
 interface Props {
   post: ForumPost
@@ -15,9 +19,30 @@ interface Props {
 }
 
 export const PostEditor = ({ post, onCancel, type }: Props) => {
+  const { api, connectionState } = useApi()
   const [newText, setNewText] = useState(post.text)
   const { showModal } = useModal()
   const isTextChanged = useMemo(() => post.text !== newText, [newText])
+
+  const forumPostData = useForumPostParents(type === 'forum' ? post.id : '')
+  const proposalPostData = useProposalPostParents(type === 'proposal' ? post.id : '')
+
+  const editPostTransaction = useMemo(() => {
+    if (api && connectionState === 'connected') {
+      if (type === 'forum' && forumPostData.categoryId && forumPostData.threadId) {
+        return api.tx.forum.editPostText(
+          createType('ForumUserId', Number.parseInt(post.author.id)),
+          forumPostData.categoryId,
+          forumPostData.threadId,
+          post.id,
+          newText
+        )
+      }
+      if (type === 'proposal' && proposalPostData.threadId) {
+        return api.tx.proposalsDiscussion.updatePost(post.id, proposalPostData.threadId, newText)
+      }
+    }
+  }, [api, connectionState, JSON.stringify(forumPostData), JSON.stringify(proposalPostData), type])
 
   return (
     <EditorWrap>
@@ -28,7 +53,12 @@ export const PostEditor = ({ post, onCancel, type }: Props) => {
         </ButtonGhost>
         <ButtonPrimary
           size="medium"
-          onClick={() => showModal<EditPostModalCall>({ modal: 'EditPost', data: { newText, post, type } })}
+          onClick={() =>
+            showModal<EditPostModalCall>({
+              modal: 'EditPost',
+              data: { post, transaction: editPostTransaction },
+            })
+          }
           disabled={!isTextChanged}
         >
           Save
