+++ packages/ui/src/forum/modals/PostActionModal/EditPostModal/index.ts
@@ -1,9 +1,11 @@
+import { SubmittableExtrinsic } from '@polkadot/api/types'
+import { ISubmittableResult } from '@polkadot/types/types'
+
 import { ModalWithDataCall } from '@/common/providers/modal/types'
-import { PostListItemType } from '@/forum/components/PostList/PostListItem'
 import { ForumPost } from '@/forum/types/ForumPost'
 
 export * from './EditPostModal'
 export type EditPostModalCall = ModalWithDataCall<
   'EditPost',
-  { post: ForumPost; newText: string; type: PostListItemType }
+  { post: ForumPost; transaction?: SubmittableExtrinsic<'rxjs', ISubmittableResult> }
 >
